requirements.txt
streamlit
openai
requests
load_dotenv
openai
PyMuPDF
Image

# 1. Install the requirements

#    ```
#    $ pip install -r requirements.txt
#    ```

# 2. Run the app

#    ```
#    $ streamlit run streamlit_app.py
#    ```
